Renamed isomorphic modules
