+++ NavigationReact/sample/isomorphic/getStateNavigator.js
