+++ NavigationReact/sample/isomorphic/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { NavigationContext, NavigationHandler } from 'navigation-react';
-import getStateNavigator from './NavigationShared';
+import getStateNavigator from './getStateNavigator';
 
 var stateNavigator = getStateNavigator();
 registerControllers(stateNavigator);
