+++ NavigationReact/sample/isomorphic/server.js
@@ -2,7 +2,7 @@ import express from 'express';
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 import { NavigationContext, NavigationHandler } from 'navigation-react';
-import getStateNavigator from './NavigationShared';
+import getStateNavigator from './getStateNavigator';
 import { searchPeople, getPerson } from './Data';
 
 var app = express();
