+++ NavigationReact/sample/isomorphic/webpack.config.js
@@ -1,7 +1,7 @@
 var path = require('path');
 
 module.exports = {
-    entry: "./NavigationClient.js",
+    entry: "./index.js",
     output: {
         path: path.join(__dirname, "js"),
         filename: "app.js"
